The arm guys have now screwed up the ARM versatile board's IRQ routing three consecutive times. I'm impressed.
The ARM versatile scsi IRQ is position 27 on the IRQ controller. That's what
QEMU implemented, that's what worked for years. In commit 1bc39ac5dab2 the
kernel guys screwed up their math (among other things doing -24 and then &3
on the result.. can we say NOP?) so it was now trying to use IRQ 28 once
the unnecessary "swizzle" function got done with it. (I started reverting
that 6 months ago in aboriginal changeset 1534.) Then in commit f5565295892e
they incremented the IRQ controller start by 32... and didn't adjust map_irq()
so it was still requesting 28, now before the start of the IRQ
controller's range. Then in commit e3e92a7be693 they noticed it was
broken, and added 64 to it. (So now it's requesting 64+28=92, when it
_should_ be requesting 32+27=59. Their own description of what changed does
not support what the patch did, and yet...)

Of _course_ they didn't test the result and notice it doesn't work. Even
though qemu is there, and worked for years before they started messing
with it.